# openapi.yaml
openapi: 3.0.3
info:
  title: FakeAPI Store
  version: 1.2.3
  description: |-
    This is a sample FakeAPI Store Server based on the OpenAPI 3.X specification. You can find out more about Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach. You can now help us improve the API whether it's by making changes to the definition itself or to the code. That way, with time, we can improve the API in general, and expose some of the new features in OAS3.
    
    Some useful links:
    - [The Fake Store repository](https://github.com/platzi/curso-apifirst)
  termsOfService: https://platzi.com/terms/
  contact:
    name: Fabian Sastoque
    url: https://platzi.com
    email: fsastoque@gmail.com
  license:
    name: MIT
    url: https://opensource.org/license/MIT
externalDocs:
  description: Find out more about FakeAPI Store
  url: https://www.panamericana.com
tags:
  - name: Hello
    description: Endpoints relacionados con el saludo
  - name: Users
    description: Endpoints relacionados con los usuarios
  - name: Products
    description: Endpoints relacionados con los productos
    externalDocs:
      description: Find out more about OpenAPI
      url: https://openai.com
servers:  
  - url: http://localhost:3000
    description: Version 1 de la API
  - url:  http://localhost:3000/v2
    description: Version 2 de la API
components:
  schemas:
    HelloResponse:
      type: object
      example:
        message: "¡Bienvenido a FakeAPI Store!"
    Product:
      type: object
      required:
        - name
        - price
        - category
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 40
          example: "Smartphone Samsung Galaxy S23"
        description:
          type: string
          maxLength: 500
          example: "Teléfono inteligente con pantalla AMOLED de 6.1'', 128GB de almacenamiento y cámara triple de 50MP."
        price:
          type: number
          minimum: 0
          multipleOf: 0.01
          example: 3499.99
        category:
          type: string
          enum: [Electronics, Clothing, Books, Other]
          example: "Electronics"
        tags:
          type: array
          items:
            type: string
            minLength: 1
          example: ["Samsung", "Smartphone", "Android"]
        inStock:
          type: boolean
          example: true
        specifications:
          type: object
          additionalProperties:
            type: string
          example:
            color: "Negro"
            garantia: "1 año"
            modelo: "SM-S911B"
        ratings:
          type: array
          items:
            type: object
            required:
              - score
              - comment
            properties:
              score:
                type: integer
                minimum: 1
                maximum: 5
                example: 5
              comment:
                type: string
                maxLength: 200
                example: "Excelente teléfono, muy rápido y con buena cámara."
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
security:
  - JWT: []
paths:
  /hello:
    get:
      tags:
        - Hello
      summary: Hello Wolrd
      responses:
        '200':
          description: A successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelloResponse'
              example:
                message: "¡Bienvenido a FakeAPI Store!"
  /:
    get:
      summary: Inicio
      responses:
        '200':
          description: A successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HelloResponse'
              example:
                message: "¡Bienvenido a FakeAPI Store!"
  /users:
    post:
      tags:
        - Users
      summary: Crea un nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - age
                - email
              properties:
                name:
                  type: string
                  minLength: 2
                  example: "Juan Pérez"
                age:
                  type: integer
                  minimum: 18
                  example: 28
                email:
                  type: string
                  format: email
                  example: "juan.perez@email.com"
            example:
              name: "Juan Pérez"
              age: 28
              email: "juan.perez@email.com"
      responses:
        '201':
          description: Usuario creado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "usr_12345"
                  name:
                    type: string
                    example: "Juan Pérez"
                  age:
                    type: integer
                    example: 28
                  email:
                    type: string
                    example: "juan.perez@email.com"
  /users/{id}:
    get:
      tags:
        - Users
      summary: Obtener usuario por Id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 12345
      responses:
        '200':
          description: Usuario encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 12345
                  name:
                    type: string
                    example: "Juan Pérez"
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Usuario no encontrado"
    post:
      tags:
        - Users
      summary: Actualizar un nuevo usuario
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 12345
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - age
                - email
              properties:
                name:
                  type: string
                  minLength: 2
                  example: "Juan Pérez"
                age:
                  type: integer
                  minimum: 18
                  example: 29
                email:
                  type: string
                  format: email
                  example: "juan.perez@email.com"
            example:
              name: "Juan Pérez"
              age: 29
              email: "juan.perez@email.com"
      responses:
        '200':
          description: Usuario actulizado correctamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: "usr_12345"
                  name:
                    type: string
                    example: "Juan Pérez"
                  age:
                    type: integer
                    example: 29
                  email:
                    type: string
                    example: "juan.perez@email.com"
        '404':
          description: Usuario no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Usuario no encontrado"
  /products:
    post:
      tags:
        - Products
      summary: Crear un nuevo producto
      security:
        - JWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
            example:
              name: "Smart TV LG 55'' 4K"
              description: "Televisor inteligente con resolución 4K UHD, compatible con Alexa y Google Assistant."
              price: 2499.99
              category: "Electronics"
              tags: ["LG", "Smart TV", "4K"]
              inStock: true
              specifications:
                color: "Negro"
                modelo: "55UQ8000"
                garantia: "2 años"
              ratings:
                - score: 5
                  comment: "Imagen espectacular y fácil de usar."
                - score: 4
                  comment: "Buen producto, pero el sonido podría mejorar."
      responses:
        '201':
          description: Producto creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              example:
                name: "Smart TV LG 55'' 4K"
                description: "Televisor inteligente con resolución 4K UHD, compatible con Alexa y Google Assistant."
                price: 2499.99
                category: "Electronics"
                tags: ["LG", "Smart TV", "4K"]
                inStock: true
                specifications:
                  color: "Negro"
                  modelo: "55UQ8000"
                  garantia: "2 años"
                ratings:
                  - score: 5
                    comment: "Imagen espectacular y fácil de usar."
                  - score: 4
                    comment: "Buen producto, pero el sonido podría mejorar."
        '400':
          description: Error de validación
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "El campo 'name' es obligatorio."
    get:
      tags:
        - Products
      summary: Listar todos los productos
      responses:
        '200':
          description: Lista de productos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
              example:
                - name: "Smartphone Samsung Galaxy S23"
                  description: "Teléfono inteligente con pantalla AMOLED de 6.1'', 128GB de almacenamiento y cámara triple de 50MP."
                  price: 3499.99
                  category: "Electronics"
                  tags: ["Samsung", "Smartphone", "Android"]
                  inStock: true
                  specifications:
                    color: "Negro"
                    garantia: "1 año"
                    modelo: "SM-S911B"
                  ratings:
                    - score: 5
                      comment: "Excelente teléfono, muy rápido y con buena cámara."
                - name: "Camisa Polo Ralph Lauren"
                  description: "Camisa polo de algodón, color azul marino, talla M."
                  price: 399.00
                  category: "Clothing"
                  tags: ["Ralph Lauren", "Camisa", "Polo"]
                  inStock: true
                  specifications:
                    talla: "M"
                    color: "Azul marino"
                  ratings:
                    - score: 4
                      comment: "Muy cómoda y de buena calidad."
  /products/{id}:
    get:
      tags:
        - Products
      summary: Obtener producto por ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: "prd_001"
      responses:
        '200':
          description: Producto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              example:
                name: "Libro: El Principito"
                description: "Edición especial ilustrada, tapa dura."
                price: 129.00
                category: "Books"
                tags: ["Libro", "Clásico", "Infantil"]
                inStock: true
                specifications:
                  autor: "Antoine de Saint-Exupéry"
                  idioma: "Español"
                  paginas: "96"
                ratings:
                  - score: 5
                    comment: "Hermosa edición, ideal para regalar."
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Producto no encontrado"
    put:
      tags:
        - Products
      summary: Actualizar producto por ID
      security:
        - JWT: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: "prd_001"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
            example:
              name: "Libro: El Principito"
              description: "Edición especial ilustrada, tapa dura."
              price: 129.00
              category: "Books"
              tags: ["Libro", "Clásico", "Infantil"]
              inStock: true
              specifications:
                autor: "Antoine de Saint-Exupéry"
                idioma: "Español"
                paginas: "96"
              ratings:
                - score: 5
                  comment: "Hermosa edición, ideal para regalar."
      responses:
        '200':
          description: Producto actualizado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
              example:
                name: "Libro: El Principito"
                description: "Edición especial ilustrada, tapa dura."
                price: 129.00
                category: "Books"
                tags: ["Libro", "Clásico", "Infantil"]
                inStock: true
                specifications:
                  autor: "Antoine de Saint-Exupéry"
                  idioma: "Español"
                  paginas: "96"
                ratings:
                  - score: 5
                    comment: "Hermosa edición, ideal para regalar."
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Producto no encontrado"
    delete:
      tags:
        - Products
      summary: Eliminar producto por ID
      security:
        - JWT: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          example: "prd_001"
      responses:
        '204':
          description: Producto eliminado correctamente
        '404':
          description: Producto no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Producto no encontrado"
  /auth/login:
    post:
      tags:
        - Users
      summary: Autenticar usuario y obtener JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "juan.perez@email.com"
                password:
                  type: string
                  format: password
                  example: "MiContraseñaSegura123"
            example:
              email: "juan.perez@email.com"
              password: "MiContraseñaSegura123"
      responses:
        '200':
          description: Autenticación exitosa, retorna el token JWT
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Credenciales inválidas"
 